fix: interrupts in the linux rt_clock.hpp
- Volatile bool interrupt was removed as it is not required
for the Linux rt_clock. The interrupt is set by the observer
pattern.
- The interrupt is now initialized in the rt_clock constructor.
It was never initialized as it was using a volatile global variable.
Which is not required for the Linux rt_clock.
- The interrupt variable is now cleared after an interrupted is
registered and handled by the set_timeout function. It was never
cleared in the Linux rt_clock.
- Changed bool interrupt in dynamic_model.hpp AsyncEventObserver
to volatile bool interrupt. This is so that the wait loop in rt_clock
does not get optimized. This is useful if the memory is changed
from another location or a callback is used with the update function.